+++ src/types/components.ts
@@ -10,6 +10,7 @@ import {
 } from 'react-hook-form';
 
 import {RegionType} from '@/types/response';
+import {Location} from '@/types/store';
 
 type InputType = 'text' | 'password' | 'email' | 'date' | 'number';
 type ButtonType = 'button' | 'submit' | 'reset';
@@ -145,3 +146,11 @@ export interface TimeInputProps {
 
 export type LatLngLiteral = google.maps.LatLngLiteral;
 export type MapOptions = google.maps.MapOptions;
+
+export interface DndCardProps {
+  items: Location[];
+  updateItem: (_item: Location[]) => void;
+  id: number;
+  accept: string;
+  index?: number;
+}
